[WIP] Scripted tor identity on captcha
This is an attempt at a scripted tor identity "reload" for a new IP,
when we see the captcha page (i.e., get redirected). However it does not
yet work. It seems, that we somehow still get redirected although the
system reports that we got a new IP address already. Also, using just
one concurrent request (to exclude blocks for connections that are
still open and "use" the old ip address) does not help. Waiting for
30 secs does not help either. The only solution is to restart the
spider (therefore the spider persists its state in the 'crawl' dir).